+++ src/js/components/lightbox.js
@@ -1,5 +1,5 @@
 import LightboxPanel from './lightbox-panel';
-import {$$, assign, data, findIndex, on, uniqueBy} from 'uikit-util';
+import {$$, assign, data, findIndex, isElement, on, uniqueBy} from 'uikit-util';
 
 export default {
 
@@ -21,10 +21,6 @@ export default {
                 this.hide();
             }
 
-        },
-
-        items() {
-            return uniqueBy(this.toggles.map(toItem), 'source');
         }
 
     },
@@ -45,8 +41,7 @@ export default {
 
             handler(e) {
                 e.preventDefault();
-                const src = data(e.current, 'href');
-                this.show(findIndex(this.items, ({source}) => source === src));
+                this.show(e.current);
             }
 
         }
@@ -57,7 +52,14 @@ export default {
 
         show(index) {
 
-            this.panel = this.panel || this.$create('lightboxPanel', assign({}, this.$props, {items: this.items}));
+            const items = uniqueBy(this.toggles.map(toItem), 'source');
+
+            if (isElement(index)) {
+                const {source} = toItem(index);
+                index = findIndex(items, ({source: src}) => source === src);
+            }
+
+            this.panel = this.panel || this.$create('lightboxPanel', assign({}, this.$props, {items}));
 
             on(this.panel.$el, 'hidden', () => this.panel = false);